refactor inference q8 runtime plan
